I-4-contributor-guidelines-and-license
This PR Adds the contributor guidelines and the MIT license. We've talked about and hashed out our opinions on contributing and everyone agrees to follow and enforce these guidelines throughout the project.

The MIT license was chosen because it's passive and permissive while still covering liability and the appropriate warranties.